--- test/integration/config-schema-check/next.config.js
@@ -1,3 +1,0 @@
-module.exports = (phase, { defaultConfig }) => {
-  return defaultConfig
-}
