--- test/integration/config-schema-check/pages/index.js
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <p>hello world</p>
-}
